wielowątkowe uczenie, biasy, zmiana z float na double, porządek w mainie

+++ src/boltzmann/layers/LayerConnectorWeightInitializer.java
@@ -2,5 +2,5 @@ package boltzmann.layers;
 
 public abstract class LayerConnectorWeightInitializer {
 
-	public abstract float getWeight();
+	public abstract double getWeight();
 }
